refactor: LiquorResultCode 패키지를 core로 이동
